+++ dlib/external/libjpeg/jddctmgr.cpp
@@ -182,7 +182,7 @@ start_pass (j_decompress_ptr cinfo)
 	 */
 	IFAST_MULT_TYPE * ifmtbl = (IFAST_MULT_TYPE *) compptr->dct_table;
 #define CONST_BITS 14
-	static const INT16 aanscales[DCTSIZE2] = {
+	static const short aanscales[DCTSIZE2] = {
 	  /* precomputed values scaled up by 14 bits */
 	  16384, 22725, 21407, 19266, 16384, 12873,  8867,  4520,
 	  22725, 31521, 29692, 26722, 22725, 17855, 12299,  6270,
@@ -197,8 +197,8 @@ start_pass (j_decompress_ptr cinfo)
 
 	for (i = 0; i < DCTSIZE2; i++) {
 	  ifmtbl[i] = (IFAST_MULT_TYPE)
-	    DESCALE(MULTIPLY16V16((INT32) qtbl->quantval[i],
-				  (INT32) aanscales[i]),
+	    DESCALE(MULTIPLY16V16((long) qtbl->quantval[i],
+				  (long) aanscales[i]),
 		    CONST_BITS-IFAST_SCALE_BITS);
 	}
       }
